fix: details and exit button
